refactor(core): use usize for block_ids, validate at RPC boundary (#351)
## Summary

`block_id` 是数组索引，但之前在 engine 内部以 `i32` 流转，并在 `lib.rs` 的 layers × blocks
嵌套循环里逐 block 做 `usize::try_from`。本 PR 把这种"不依赖 PegaEngine
状态"的参数校验与类型转换统一上提到 RPC service 边界，engine 内部直接讲 `usize`。

按"值的语义"而非"来源"区分类型：

- **索引 / 计数**(`block_ids`)→ engine 内部统一 `usize`。`i32`/`u32` 只是 protobuf
没有 `usize` 的 wire 产物,应只活在 RPC 边界。
- **`device_id`** 保持 `i32`:它是 CUDA device handle(直接传给 `CudaContext::new`
/ `numa_for_gpu`,CUDA 用 signed int),不是索引,强转 `usize` 只会在每个 CUDA 调用处再 `as
i32`。
- **PyO3 gRPC client** 保持 `Vec<i32>`:那是 protobuf wire 类型,边界本就在那里。

## Changes

- `pegaflow-core`: load API(`batch_load_kv_blocks_multi_layer` /
`_inproc` / `_inner`)与 `LayerSave.block_ids` 改为 `Vec<usize>`;移除 `lib.rs`
嵌套循环里的逐 block `usize::try_from` 与 `offload.rs` 的 `id as usize`。
- `pegaflow-server/service.rs`:`i32 -> usize` 转换(含拒绝负值)在 `load` / `save`
handler 各做一次,复用既有标量校验风格;`validate_save_layers` 里冗余的负值检查删除。
- 同步更新 test / bench / example 调用点为 `Vec<usize>`。

## Testing

- `cargo build` / `cargo clippy --tests --benches --examples`:clean。
- `cargo test --release -p pegaflow-core --no-default-features
--features cuda-13,rdma`:全绿(含 `save_validation`、`prefetch_lease`)。
- `cargo test --release -p pegaflow-server --no-default-features
--features cuda-13,rdma`:26 个单测全绿(含 `service.rs` 校验路径)。
- `http_cleanup_hang_repro` 在本机崩溃(torch CUDA/GIL interpreter 冲突),已用 `git
stash` 在干净 master 上复现同样崩溃,**与本改动无关**。

> ⚠️ 本机默认 feature 是 `cuda-12`,但运行时是 CUDA 13 driver,需用 `--features
cuda-13` 跑 GPU 相关测试。Reviewer 请在 cu13 机器上复跑 e2e gate。

## Breaking change

`batch_load_kv_blocks_multi_layer(_inproc)` 与 `LayerSave.block_ids` 现在接受
`usize` 而非 `i32`。直接调用这些 engine API 的 in-process Rust embedder 需要相应改为传
`usize`。gRPC / proto wire 协议**不变**,Python 侧无感知。

+++ pegaflow-core/src/block.rs
@@ -21,7 +21,7 @@ pub type BlockHash = Vec<u8>;
 /// Per-layer save input: layer name + block IDs + content hashes.
 pub struct LayerSave {
     pub layer_name: String,
-    pub block_ids: Vec<i32>,
+    pub block_ids: Vec<usize>,
     pub block_hashes: Vec<Vec<u8>>,
 }
